refactor: change route configuration for profile page into dashboard.
All file in the profile page moved into the newly created dashboarc file.
Move the file in the initila profile/Dashboard route into the new dashboard file without any subfolder nesting
fix all import for moved files.
move the Sidebar component in the initial pages/profile  folder into the component/profile folder
update the route configuration in the utils/config file.
