+++ src/pages/dashboard/purchases.js
@@ -1,0 +1,9 @@
+import ProfileLayout from "../../layout/profileLayout";
+
+const Purchases = () => {
+  return <div>Purchases</div>;
+};
+
+Purchases.PageLayout = ProfileLayout;
+
+export default Purchases;
